Extract a contract form in a dedicated file
