+++ assets/javascripts/application.js
@@ -10,7 +10,7 @@ import CheckboxesController from '@/controllers/checkboxes_controller.js';
 import ColorSchemeController from '@/controllers/color_scheme_controller.js';
 import EditorController from '@/controllers/editor_controller.js';
 import FormNewAuthorizationController from '@/controllers/form_new_authorization_controller.js';
-import FormNewContractController from '@/controllers/form_new_contract_controller.js';
+import FormContractController from '@/controllers/form_contract_controller.js';
 import FormPriorityController from '@/controllers/form_priority_controller.js';
 import MessageDocumentsController from '@/controllers/message_documents_controller.js';
 import ModalController from '@/controllers/modal_controller.js';
@@ -30,7 +30,7 @@ application.register('checkboxes', CheckboxesController);
 application.register('color-scheme', ColorSchemeController);
 application.register('editor', EditorController);
 application.register('form-new-authorization', FormNewAuthorizationController);
-application.register('form-new-contract', FormNewContractController);
+application.register('form-contract', FormContractController);
 application.register('form-priority', FormPriorityController);
 application.register('message-documents', MessageDocumentsController);
 application.register('modal', ModalController);
