+++ assets/javascripts/controllers/form_contract_controller.js
