+++ templates/organizations/contracts/new.html.twig
@@ -20,163 +20,9 @@
             <h1>{{ 'contracts.new.title' | trans }}</h1>
         </div>
 
-        {{ form_start(form, { attr: {
-            'class': 'wrapper wrapper--small wrapper--center flow',
-            'data-controller': 'form-new-contract',
-        }}) }}
-            {{ form_errors(form) }}
-
-            <div class="flow flow--small">
-                <label for="{{ field_id(form.name) }}">
-                    {{ 'contracts.name' | trans }}
-
-                    <small class="text--secondary">
-                        {{ 'forms.max_chars' | trans({ number: 255 }) }}
-                    </small>
-                </label>
-
-                {{ form_errors(form.name) }}
-
-                <input
-                    type="text"
-                    id="{{ field_id(form.name) }}"
-                    name="{{ field_name(form.name) }}"
-                    value="{{ field_value(form.name) }}"
-                    required
-                    maxlength="255"
-                    {% if field_has_errors(form.name) %}
-                        aria-invalid="true"
-                        aria-errormessage="{{ field_id(form.name, 'error') }}"
-                    {% endif %}
-                />
-            </div>
-
-            <div class="row flow">
-                <div class="row__item--extend flow flow--small">
-                    <label for="{{ field_id(form.startAt) }}">
-                        {{ 'contracts.start_at' | trans }}
-                    </label>
-
-                    {{ form_errors(form.startAt) }}
-
-                    <input
-                        type="date"
-                        id="{{ field_id(form.startAt) }}"
-                        name="{{ field_name(form.startAt) }}"
-                        value="{{ field_value(form.startAt) }}"
-                        required
-                        data-form-new-contract-target="startAt"
-                        data-action="form-new-contract#updateEndAt"
-                        {% if field_has_errors(form.startAt) %}
-                            aria-invalid="true"
-                            aria-errormessage="{{ field_id(form.startAt, 'error') }}"
-                        {% endif %}
-                    />
-                </div>
-
-                <div class="row__item--extend flow flow--small">
-                    <label for="{{ field_id(form.endAt) }}">
-                        {{ 'contracts.end_at' | trans }}
-                    </label>
-
-                    {{ form_errors(form.endAt) }}
-
-                    <input
-                        type="date"
-                        id="{{ field_id(form.endAt) }}"
-                        name="{{ field_name(form.endAt) }}"
-                        value="{{ field_value(form.endAt) }}"
-                        required
-                        data-form-new-contract-target="endAt"
-                        {% if field_has_errors(form.endAt) %}
-                            aria-invalid="true"
-                            aria-errormessage="{{ field_id(form.endAt, 'error') }}"
-                        {% endif %}
-                    />
-                </div>
-            </div>
-
-            <div class="flow flow--small">
-                <label for="{{ field_id(form.maxHours) }}">
-                    {{ 'contracts.max_hours' | trans }}
-                </label>
-
-                {{ form_errors(form.maxHours) }}
-
-                <input
-                    type="number"
-                    id="{{ field_id(form.maxHours) }}"
-                    name="{{ field_name(form.maxHours) }}"
-                    value="{{ field_value(form.maxHours) }}"
-                    min="0"
-                    step="1"
-                    required
-                    class="input--size3"
-                    {% if field_has_errors(form.maxHours) %}
-                        aria-invalid="true"
-                        aria-errormessage="{{ field_id(form.maxHours, 'error') }}"
-                    {% endif %}
-                />
-            </div>
-
-            <div class="flow flow--small">
-                <label for="{{ field_id(form.billingInterval) }}">
-                    {{ 'contracts.new.billing_interval' | trans }}
-
-                    <small class="text--secondary">
-                        {{ 'forms.optional' | trans }}
-                    </small>
-                </label>
-
-                <p class="form__caption" id="{{ field_id(form.billingInterval, 'caption') }}">
-                    {{ 'contracts.new.billing_interval.caption' | trans }}
-                </p>
-
-                {{ form_errors(form.billingInterval) }}
-
-                <input
-                    type="number"
-                    id="{{ field_id(form.billingInterval) }}"
-                    name="{{ field_name(form.billingInterval) }}"
-                    value="{{ field_value(form.billingInterval) }}"
-                    min="0"
-                    step="1"
-                    class="input--size3"
-                    aria-describedby="{{ field_id(form.billingInterval, 'caption') }}"
-                    {% if field_has_errors(form.billingInterval) %}
-                        aria-invalid="true"
-                        aria-errormessage="{{ field_id(form.billingInterval, 'error') }}"
-                    {% endif %}
-                />
-            </div>
-
-            <div class="flow flow--small">
-                <label for="{{ field_id(form.notes) }}">
-                    {{ icon('user-secret') }}
-                    {{ 'contracts.notes' | trans }}
-
-                    <small class="text--secondary">
-                        {{ 'forms.optional' | trans }}
-                    </small>
-                </label>
-
-                {{ form_errors(form.notes) }}
-
-                <textarea
-                    id="{{ field_id(form.notes) }}"
-                    name="{{ field_name(form.notes) }}"
-                    {% if field_has_errors(form.notes) %}
-                        aria-invalid="true"
-                        aria-errormessage="{{ field_id(form.notes, 'error') }}"
-                    {% endif %}
-                >{{ field_value(form.notes) }}</textarea>
-            </div>
-
-            <div class="form__actions">
-                <button class="button--primary" type="submit">
-                    {{ 'contracts.new.submit' | trans }}
-                </button>
-            </div>
-        {{ form_end(form) }}
+        {{ include('contracts/_form.html.twig', {
+            'form': form,
+            'submit_label': 'contracts.form.new' | trans,
+        }) }}
     </main>
 {% endblock %}
