+++ translations/messages+intl-icu.en_GB.yaml
@@ -53,7 +53,14 @@ contracts.alerts.edit.hours_alert.after: '% of the hours have been consumed.'
 contracts.alerts.edit.hours_alert.before: 'Alert when'
 contracts.alerts.edit.title: 'Set up contract alerts'
 contracts.days_consumed: "<strong>{days}\_d consumed</strong>"
-contracts.end_at: 'End on'
+contracts.form.billing_interval: 'Billing interval in minutes'
+contracts.form.billing_interval.caption: 'The time spent on tickets will be rounded up for billing purposes (e.g. for 15 minutes of time spent, 30 minutes will be deducted for a corresponding billing interval).'
+contracts.form.end_at: 'End on'
+contracts.form.max_hours: 'Max number of hours'
+contracts.form.name: Name
+contracts.form.new: 'Create the contract'
+contracts.form.notes: 'Private notes'
+contracts.form.start_at: 'Start on'
 contracts.hours_consumed: '<strong>{hours} consumed</strong> on {maxHours}h'
 contracts.index.alert: Alert
 contracts.index.new_contract: 'New contract'
@@ -61,13 +68,8 @@ contracts.index.no_contracts: 'No contracts'
 contracts.index.number: '{count, plural, =0 {No contracts} one {1 contract} other {# contracts}}'
 contracts.index.tickets: '{count, plural, =0 {No tickets} one {1 ticket} other {# tickets}}'
 contracts.index.title: Contracts
-contracts.max_hours: 'Max number of hours'
 contracts.max_hours.edit.additional_hours: 'Number of additional hours'
 contracts.max_hours.edit.title: 'Extend the number of hours of the contract'
-contracts.name: Name
-contracts.new.billing_interval: 'Billing interval in minutes'
-contracts.new.billing_interval.caption: 'The time spent on tickets will be rounded up for billing purposes (e.g. for 15 minutes of time spent, 30 minutes will be deducted for a corresponding billing interval).'
-contracts.new.submit: 'Create the contract'
 contracts.new.title: 'New contract'
 contracts.notes: 'Private notes'
 contracts.show.alert: Alert
@@ -77,7 +79,6 @@ contracts.show.notes_confidential: confidential
 contracts.show.renew: Renew
 contracts.show.set_up_alerts: 'Set up alerts'
 contracts.show.warning: Warning
-contracts.start_at: 'Start on'
 contracts.status: Status
 contracts.status.coming: Coming
 contracts.status.finished: Finished
